Remove check.{hpp,cpp}
- Remove check.hpp and check.cpp - merge into evaluate.hpp and evaluate.cpp
- In evaluate(), return early when checkmate found to avoid modifying it with castle penalty. Shouldn't be a problem in most cases due to current values anyway, but we shouldn't modify the score after checkmate is detected
- Update comments
