+++ engine/src/evaluate.cpp
@@ -1,6 +1,5 @@
 #include "evaluate.hpp"
 #include "board.hpp"
-#include "check.hpp"
 #include "position.hpp"
 #include "search.hpp"
 
@@ -8,53 +7,50 @@ namespace wisdom
 {
     static constexpr int Castle_Penalty = 50;
 
-    namespace
+    [[nodiscard]] static auto
+    heuristicIsCastled (const Board& board, Color who)
+        -> bool
     {
-        auto 
-        heuristicIsCastled (const Board& board, Color who) 
-            -> bool
-        {
-            auto king_pos = board.getKingPosition (who);
-            auto king_column = king_pos.column<int>();
-            auto king_row = king_pos.row<int>();
+        auto king_pos = board.getKingPosition (who);
+        auto king_column = king_pos.column<int>();
+        auto king_row = king_pos.row<int>();
 
-            auto rook_piece = ColoredPiece::make (who, Piece::Rook);
+        auto rook_piece = ColoredPiece::make (who, Piece::Rook);
 
-            if (king_row != castlingRowForColor (who))
-                return false;
-
-            // check kingside castle:
-            if (king_column == Kingside_Castled_King_Column)
-            {
-                if (board.pieceAt (king_row, Kingside_Castled_Rook_Column) == rook_piece)
-                    return true;
-            }
-            else if (king_column == Queenside_Castled_King_Column)
-            {
-                if (board.pieceAt (king_row, Queenside_Castled_Rook_Column) == rook_piece)
-                    return true;
-            }
-
+        if (king_row != castlingRowForColor (who))
             return false;
+
+        // check kingside castle:
+        if (king_column == Kingside_Castled_King_Column)
+        {
+            if (board.pieceAt (king_row, Kingside_Castled_Rook_Column) == rook_piece)
+                return true;
+        }
+        else if (king_column == Queenside_Castled_King_Column)
+        {
+            if (board.pieceAt (king_row, Queenside_Castled_Rook_Column) == rook_piece)
+                return true;
         }
 
-        auto 
-        unableToCastlePenalty (const Board& board, Color who)
-            -> int
+        return false;
+    }
+
+    [[nodiscard]] static auto
+    unableToCastlePenalty (const Board& board, Color who)
+        -> int
+    {
+        auto castle_state = board.getCastlingEligibility (who);
+        int result = 0;
+        if (castle_state != Either_Side_Eligible)
         {
-            auto castle_state = board.getCastlingEligibility (who);
-            int result = 0;
-            if (castle_state != Either_Side_Eligible)
-            {
-                if (castle_state & CastlingIneligible::Kingside)
-                    result += Castle_Penalty;
-                if (castle_state & CastlingIneligible::Queenside)
-                    result += Castle_Penalty;
-                if (heuristicIsCastled (board, who))
-                    result -= 2 * Castle_Penalty;
-            }
-            return result;
+            if (castle_state & CastlingIneligible::Kingside)
+                result += Castle_Penalty;
+            if (castle_state & CastlingIneligible::Queenside)
+                result += Castle_Penalty;
+            if (heuristicIsCastled (board, who))
+                result -= 2 * Castle_Penalty;
         }
+        return result;
     }
 
     auto 
@@ -64,17 +60,17 @@ namespace wisdom
         int score = 0;
         Color opponent = colorInvert (who);
 
-        score += board.getMaterial().overallScore (who);
-        score += board.getPosition().overallScore (who);
-
         if (isCheckmated (board, who))
         {
-            score = -1 * checkmateScoreInMoves (moves_away);
+            return -1 * checkmateScoreInMoves (moves_away);
         }
         else if (isCheckmated (board, opponent))
         {
-            score = checkmateScoreInMoves (moves_away);
+            return checkmateScoreInMoves (moves_away);
         }
+
+        score += board.getMaterial().overallScore (who);
+        score += board.getPosition().overallScore (who);
 
         score -= unableToCastlePenalty (board, who);
         score += unableToCastlePenalty (board, opponent);
@@ -87,8 +83,61 @@ namespace wisdom
         -> int
     {
         auto king_coord = board.getKingPosition (who);
-        return isKingThreatened (board, who, king_coord)
-                       ? -1 * checkmateScoreInMoves (moves_away)
-                       : 0;
+        return isKingThreatened (board, who, king_coord) 
+            ? -1 * checkmateScoreInMoves (moves_away) 
+            : 0;
+    }
+
+    bool isCheckmated (const Board& board, Color who)
+    {
+        auto coord = board.getKingPosition (who);
+
+        if (!isKingThreatened (board, who, coord))
+            return false;
+
+        MoveList legal_moves = generateLegalMoves (board, who);
+
+        return legal_moves.isEmpty();
+    }
+
+    auto
+    isLegalPositionAfterMove (const Board& board, Color who, Move mv)
+    -> bool
+    {
+        auto king_coord = board.getKingPosition (who);
+
+        if (isKingThreatened (board, who, king_coord))
+            return false;
+
+        if (mv.isCastling())
+        {
+            Coord castled_pos = mv.getDst();
+            auto castled_row = castled_pos.row();
+            auto castled_col = castled_pos.column();
+
+            assert (king_coord.row() == castled_row);
+            assert (king_coord.column() == castled_col);
+
+            int8_t direction = mv.isCastlingOnKingside() ? -1 : 1;
+
+            int8_t plus_one_column = nextColumn (castled_col, direction);
+            int8_t plus_two_column = nextColumn (plus_one_column, direction);
+
+            if (isKingThreatened (board, who, castled_row, plus_one_column)
+                || isKingThreatened (board, who, castled_row, plus_two_column))
+            {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    auto isStalemated (const Board& board, Color who) -> bool
+    {
+        auto coord = board.getKingPosition (who);
+        auto legal_moves = generateLegalMoves (board, who);
+
+        return legal_moves.isEmpty() && !isKingThreatened (board, who, coord);
     }
 }
